MM-25403: clean old CSS styles
When we import CSS rules from JavaScript, webpack conveniently injects an inline style tag into the `<head>` element to effect the rules in question. If a plugin is disabled or uninstalled, these styles are not cleaned up: in fact, on upgrade, there are now two (or more) sets of styles depending on when the user refreshes their browser.

In many cases, the new rules exactly match the old ones, so no bad side effects are observed. But other times, the multiplicity of rules leads to unusual rendering issues such as seen in https://mattermost.atlassian.net/browse/MM-25403. The cause there was an old `left: X` directive that was later replaced by a `margin-left` (IIRC). With both rules in play (until the next refresh), the rendering was all wrong.

To fix this, configure Webpack to apply a class namespaced to the plugin id to all injected `style` elements. Ideally, we'd just remove these on `uninitialize` and move on. But due to a regression in the webapp, `uninitialize` is not called consistently. We can't just do this on `initialize`, since by then Webpack has already injected some subset of our CSS assets.

To work around this, we add another class with a unique `BUILD_TIMESTAMP` to all rules injected by the current instance of the plugin. On startup, we look for all rules that don't match the current `BUILD_TIMESTAMP` and remove /those/.

Note that this won't fix the v0.3 -> v0.4 migration, since without the necessary tagging by class name, v0.4 can't identify the rules injected by v0.3. But going forward, it should ensure we always start with a "clean slate" and don't have to instruct users to refresh.

Longer term, this should probably get folded back into the starter-template. I'm deferring that until we fix `uninitialize` in the webapp so that the scope of the changes are minimized.

Oh, and to make the `webpack` config easier, I've switched it to TypeScript since that sanely supports ES6 modules.

Phew. Looking forward to feedback on all this!

Fixes: https://mattermost.atlassian.net/browse/MM-25403